Fix ImageTargetFileStbImage row bytes calculation for UINT8 data
The row bytes calculation was incorrectly using sizeof(float) instead of
sizeof(uint8_t) for UINT8 data types, causing incorrect buffer sizes.

+++ src/cinder/ImageTargetFileStbImage.cpp
@@ -83,7 +83,7 @@ ImageTargetFileStbImage::ImageTargetFileStbImage( DataTargetRef dataTarget, Imag
 	}
 	else {
 		setDataType( ImageIo::DataType::UINT8 );
-		mRowBytes = mNumComponents * imageSource->getWidth() * sizeof(float);
+		mRowBytes = mNumComponents * imageSource->getWidth() * sizeof(uint8_t);
 	}
 
 	if( mDataTarget->providesFilePath() ) {
